+++ internal/export/monzo.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/HallyG/fingrab/internal/api/monzo"
 	"github.com/HallyG/fingrab/internal/domain"
+	"github.com/HallyG/fingrab/internal/monzo"
 	"github.com/rs/zerolog"
 )
