+++ internal/export/starling_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/HallyG/fingrab/internal/api/starling"
 	"github.com/HallyG/fingrab/internal/domain"
 	"github.com/HallyG/fingrab/internal/export"
+	"github.com/HallyG/fingrab/internal/starling"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 )
