+++ internal/monzo/testdata/api/error.json
