+++ internal/starling/starling_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/HallyG/fingrab/internal/api/starling"
 	"github.com/HallyG/fingrab/internal/domain"
+	"github.com/HallyG/fingrab/internal/starling"
 	"github.com/HallyG/fingrab/internal/util/testutil"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
